+++ compiler_tester/src/compiler_tester/main.rs
@@ -76,30 +76,8 @@ fn main_inner(arguments: Arguments) -> anyhow::Result<()> {
         )
         .expect("Always valid");
     solx::EXECUTABLE
-        .set(
-            arguments
-                .solx
-                .unwrap_or_else(|| PathBuf::from("solx")),
-        )
+        .set(arguments.solx.unwrap_or_else(|| PathBuf::from("solx")))
         .expect("Always valid");
-
-    ///
-    /// Reads the benchmarking context from a JSON file and validates its correctness.
-    /// Benchmarking context provides additional information about benchmarking that
-    /// will be used to generate a report.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an error if
-    /// - file can't be read,
-    /// - deserialization from JSON file failed,
-    /// - the context validation failed.
-    fn read_context(path: PathBuf) -> anyhow::Result<benchmark_analyzer::BenchmarkContext> {
-        let contents = std::fs::read_to_string(path)?;
-        let context: benchmark_analyzer::BenchmarkContext = serde_json::de::from_str(&contents)?;
-        benchmark_analyzer::validate_context(&context)?;
-        Ok(context)
-    }
 
     let debug_config = if arguments.debug {
         std::fs::create_dir_all(compiler_tester::DEBUG_DIRECTORY)?;
@@ -256,7 +234,9 @@ fn main_inner(arguments: Arguments) -> anyhow::Result<()> {
 
     if let Some(path) = arguments.benchmark {
         let context = if let Some(context_path) = arguments.benchmark_context {
-            Some(read_context(context_path)?)
+            Some(benchmark_analyzer::BenchmarkContext::try_from_path(
+                context_path,
+            )?)
         } else {
             None
         };
